requirements.txt
streamlit
torch
transformers
pandas
nltk
scikit-learn
accelerate
janome
pyngrok
python-pptx
google-generativeai
